third_party/WebKit/LayoutTests/fast/frames/adopt-iframe-into-itself-expected.txt

ALERT: HierarchyRequestError: A Node was inserted somewhere it doesn't belong.
